Support Multithreading for Windows

+++ programs/lz4io.c
@@ -121,7 +121,7 @@ static void LZ4IO_finalTimeDisplay(TIME_t timeStart, clock_t cpuStart, unsigned 
         double const seconds = (double)(duration_ns + !duration_ns) / (double)1000000000.;
         double const cpuLoad_s = (double)(clock() - cpuStart) / CLOCKS_PER_SEC;
         DISPLAYLEVEL(3,"Done in %.2f s ==> %.2f MiB/s  (cpu load : %.0f%%)\n", seconds,
-                        size / seconds / 1024 / 1024,
+                        (double)size / seconds / 1024. / 1024.,
                         (cpuLoad_s / seconds) * 100.);
     }
 }
